query weather forecast report

+++ src/main/java/com/hubery/forecast/domain/GeneralWeatherReport.java
@@ -10,5 +10,5 @@ public class GeneralWeatherReport {
 
   private Integer temperature;
 
-  private Integer windSpeed;
+  private Double windSpeed;
 }
